Check GDPR API JWT "loa" in permission checks
